refactor(task): standardize DTO structure and domain types (Phase 1e)
- Create CreateTaskRequest and UpdateTaskRequest DTO stubs
- Create CreateWorkCategoryRequest and UpdateWorkCategoryRequest DTO stubs
- Centralize task type imports through the domain layer index
- Update task and work category services and mutation hooks to import from domain index
