+++ types/task/work-category-create.ts
@@ -1,0 +1,6 @@
+// TODO: Phase 3 — implement CreateWorkCategoryRequest and createWorkCategoryToJson
+export interface CreateWorkCategoryRequest {
+  name: string;
+  description?: string;
+  icon?: string;
+}
