+++ types/task/work-category-update.ts
@@ -1,0 +1,6 @@
+// TODO: Phase 3 — implement UpdateWorkCategoryRequest
+export interface UpdateWorkCategoryRequest {
+  name?: string;
+  description?: string;
+  icon?: string;
+}
